btrfs-progs: libbtrfs: revert to v6.0.2 ioctl.h
The file ioctl.h has been copied to libbtrfs in 5fc1d0cd64ee
("btrfs-progs: copy ioctl.h into libbtrfs") but that was after changes
to the btrfs_qgroup_limit in 03451430de7c ("btrfs-progs: rename qgroup
items to match the kernel naming scheme"). This caused build breakage
but hasn't been fixed in v6.1.1.

Revert the contents of libbtrfs/ioctl.h to be completely just v6.0.2
though the other changes may not be needed, it's the version we know
works.

Issue: #566
Signed-off-by: David Sterba <dsterba@suse.com>