btrfs-progs: help: add option text formatting infrastructure
To make option formatting a bit easier so the spacing is unified add
macros and formatting helpers.

Usage in the help text:

	OPTLINE("-o value", "description")

Internally the option and description are delimiters by chars that are
not part of normal text, the formatter separates that and uses fixed
with for output. The description text can be of any length, multi-line
text should still end up as one token (i.e. newline without ',' between).

Signed-off-by: David Sterba <dsterba@suse.com>